requirements.txt
streamlit
torch
transformers
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
faiss-cpu
pypdf
python-docx
fpdf
markdown
yake
sentence-transformers
numpy
pytesseract
pypdfium2
Pillow
